move input box and button into BitProcessor
Make BitProcessor a QWidget and move the input box and "Inspect" button
into BitProcessor.  This makes the current signal and slot connection
much easier and will similarly aid with future signal/slot connections
and adding new widgets, such as output textboxes and more buttons.
According to the Qt documentation, this new setup is the "right way" to
organize one's program.