add GPLv3 license in COPYING.txt and file headers
Apply the GNU General Public License version 3 and a name and description
to all code files and add the contents of the license to COPYING.txt.  The
autogen.sh file does not include the license because its contents are
trivial.

The GPL is used here to maintain compatibility with GPL versions of Qt
(those prior to version 4.5), which require that projects using the open
source version of Qt be GPL-licensed.  Some operating systems that will be
supported, such as Ubuntu 8.04, use a pre-4.5 version of Qt which is why
this licensing is needed.