add instructions and example bits labels
Add labels to the top of the layout describing what should go in the input
textbox and providing an example that the user can paste in.

Adding these labels has the side effect of making the window wider by
default.  It also lets the window maintain the same height while swapping
between input and output modes.  Previously, the height increased when
switching into output mode the first time.