set output fields even if error in bc_decode
The output fields are now set regardless of the return value of bc_decode.
This is possible because bc_decode guarantees the result track data will
be valid even if an error occurs.  Prior to this change, the output fields
remained at their previous values if an error occurred during bc_decode
rather than being set to the new partial values (if some data was decoded)
or cleared (if no data could be decoded).