add links label to input mode to help user
Add a label containing links to additional example bitstreams and a guide
on how to make your own as well as contact information for reporting
problems and suggestions.  This label is a bit too large and will probably
be hidden by default in a future version (with a toggle button).