"More Information" and "Less Information" buttons
Add buttons to toggle the display of additional information in input mode.
Unfortunately, this change causes Qt's layout engine to make the window
too small to display the additional information when it is unhidden.  As a
result, a minimum width and height were set, which force the window to be
large enough to show the additional information.  However, this disables
the automatic width resizing when a track output label exceeds the current
width of the window, causing long track outputs to appear truncated unless
the user manually resizes the window.  As a result, this solution to
hiding the additional information is unworkable because neither option is
acceptable.  The next commit will use an alternate approach to displaying
the additional information.