in input mode, make sure focus is on input textbox
Whenever the program is in input mode, ensure that the focus is on the
input textbox so that the user can immediately swipe a card or paste in
data when they start the program or click "Inspect Another".  To do this,
we set the Qt::NoFocus policy on the other selectable elements in input
mode.