combine example_bits_heading into instructions
Put the contents of example_bits_heading back into the instructions label.
This effectively reverts commit f01c7fd, which was intended to make hiding
fields easier, but turned out to be unnecessary because hiding fields
proved to be impractical.  This change causes the text displayed in input
mode to take up less space.