convert README and COPYING to Unix-style newlines
The README and COPYING text files previously used Windows-style newlines
in order to make it easier for developers on Windows to view those files.
This ease-of-use is more properly achieved by turning on the
"core.autocrlf" option in git.  Since libbitconvert already suggests this,
Bit Inspector will also suggest it when the Windows build instructions are
added.