add ../pcre to library path
In the case where a system libpcre does not exist, a ../pcre can be
created, which will be used by the build system.  This behavior is
consistent with that of libbitconvert's with respect to custom libpcre
support (see libbitconvert commit 5e85f7d).  Added a comment indicating
how ../pcre is used.  Note that INCLUDEPATH does not need modification
since Bit Inspector does not include "pcre.h" at all, it just links with
libbitconvert, which included "pcre.h" and thus requires libpcre to be
linked in.