use setFocus on input instead of NoFocus policies
Use explicit setFocus call on the input textbox during startup and on
"Inspect Another" instead of implying the focus through Qt::NoFocus
policies on the other widgets in input mode.  This is a much more
reliable way of ensuring that focus is on the input textbox at the
appropriate times.

In addition, the example_bits Qt::NoFocus policy on Windows causes
example_bits to be unhighlightable.  Text can still be selected, but it is
not highlighted, making it impossible to see which text was selected.  That
problem is solved by this change.