handle bc_find_fields errors specially
When no format matches, make the error text black so the user does not
think the error is a fatal error and add "Unknown card" prefix.  On any
other error, make the error text red and prefix with "Find fields error".
If the error is a format-related error (except for missing format file,
which is verbosely described by the bc_strerror text already), then add
"problem in formats file" to prefix.