add check and warning message for empty input
If the input was empty (detected by seeing if all output fields are NULL)
then print a warning message indicating that there's nothing to decode and
telling user how to enter something to decode.  This is more user-friendly
than the previous message the user saw in this case: "Unknown card".