bump version to 1.0 Beta 1a
Increase the version number to 1.0 Beta 1a (was previously 1.0 Beta 1).
This is done to ensure that only one revision of the repository (4ce87c3)
corresponds to 1.0 Beta 1, which is the first binary release.  This makes
it easier to see whether a given executable was built from the codebase
used for 1.0 Beta 1 just by looking at the title bar.  Intermediate
version numbers (such as 1.0 Alpha 1 and 1.0 Beta 1a) correpond to a range
of revisions in the repository.